+++ UserService/Controllers/WorkoutController.cs
@@ -35,13 +35,19 @@ namespace UserService.Controllers
         [HttpGet("{id}")]
             public async Task<IActionResult> GetById(int id)
             {
-                var workout = await _workoutService.GetByIdAsync(id);
+            var userId = int.Parse(
+             User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
 
-                if (workout == null)
-                    return NotFound();
+            var workout = await _workoutService.GetByIdAsync(id);
 
-                return Ok(workout);
-            }
+            if (workout == null)
+                return NotFound();
+
+            if (workout.UserId != userId)
+                return Forbid();
+
+            return Ok(workout);
+        }
 
         [HttpPost]
         public async Task<IActionResult> Create(CreateWorkout dto)
@@ -55,20 +61,45 @@ namespace UserService.Controllers
         }
 
         [HttpPut("{id}")]
-            public async Task<IActionResult> Update(int id, UpdateWorkout dto)
-            {
-                await _workoutService.UpdateAsync(id, dto);
+        public async Task<IActionResult> Update(int id, UpdateWorkout dto)
+        {
+            var userId = int.Parse(
+           User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
 
-                return Ok("Workout Updated Successfully");
-            }
+            var workout = await _workoutService.GetByIdAsync(id);
 
-            [HttpDelete("{id}")]
-            public async Task<IActionResult> Delete(int id)
-            {
-                await _workoutService.DeleteAsync(id);
+            if (workout == null)
+                return NotFound();
 
-                return Ok("Workout Deleted Successfully");
-            }
+            if (workout.UserId != userId)
+                return Forbid();
+
+            await _workoutService.UpdateAsync(id, dto);
+
+            return Ok("Workout Updated Successfully");
+        }
+
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> Delete(int id)
+        {
+            var userId = int.Parse(
+            User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+
+            var workout = await _workoutService.GetByIdAsync(id);
+
+            if (workout == null)
+                return NotFound();
+
+            if (workout.UserId != userId)
+                return Forbid();
+
+            await _workoutService.DeleteAsync(id);
+
+            return Ok("Workout Deleted Successfully");
+
+        }
+          
+            
         }
     }
